pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-a2a"
version = "0.6.0"
description = "Expose FastAPI applications as A2A Protocol agents — zero boilerplate, typed skills, pluggable registry, production-grade security."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["a2a", "fastapi", "agent", "llm", "registry", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.111.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "uvicorn[standard]>=0.29.0",
    "tenacity>=8.3.0",
    "anyio>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]
vector-weaviate = ["weaviate-client>=4.5.0"]
vector-pinecone = ["pinecone-client>=3.2.0"]
vector-qdrant = ["qdrant-client>=1.9.0"]

[tool.hatch.build.targets.wheel]
packages = ["fastapi_a2a"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B"]
ignore = ["S101", "B008", "E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = ["integration: marks tests as integration tests (require live DB)"]

[tool.coverage.run]
source = ["fastapi_a2a"]
omit = ["tests/*", "alembic/*"]
